Address failed tests

+++ sports-venue-microservice/src/application/use-cases/deleteSportsVenue.ts
@@ -8,9 +8,7 @@ import { NotFoundException } from '../../domain/exceptions/NotFoundException';
 import { UnauthorizedException } from '../../domain/exceptions/UnauthorizedException';
 import { publishSportsVenueDeletion } from '../../infrastructure/middlewares/rabbitmq.publisher';
 
-export const deleteSportsVenue = async (
-  req: Request
-): Promise<{ status: number; message: string }> => {
+export const deleteSportsVenue = async (req: Request): Promise<boolean> => {
   const id = req.params.id.toString();
   if (!mongoose.Types.ObjectId.isValid(id)) {
     throw new BadRequestException('Invalid ID format');
@@ -22,40 +20,29 @@ export const deleteSportsVenue = async (
     throw new InternalServerErrorException('Internal Server Error');
   }
 
-  try {
-    const sportsVenue = await sportsVenueRepository.findById(id);
-    if (!sportsVenue) {
-      throw new NotFoundException('Sports Venue with given ID not found');
-    }
-
-    if (sportsVenue.ownerId.toString() !== ownerId.toString()) {
-      throw new UnauthorizedException(
-        'User is not authorized to delete this venue'
-      );
-    }
-
-    await sportsVenueRepository.delete(id);
-    await publishSportsVenueDeletion({ sportsVenueId: id, ownerId });
-
-    return { status: 200, message: 'Sports Venue Deleted' };
-  } catch (error: any) {
-    switch (error) {
-      case error instanceof BadRequestException:
-        throw new BadRequestException(error.message);
-
-      case error instanceof UnauthorizedException:
-        throw new UnauthorizedException(error.message);
-
-      case error instanceof NotFoundException:
-        throw new NotFoundException(error.message);
-
-      case error instanceof ForbiddenException:
-        throw new ForbiddenException(error.message);
-
-      default:
-        throw new InternalServerErrorException(
-          'Something went wrong with sports-venue delete'
-        );
-    }
+  const sportsVenue = await sportsVenueRepository.findById(id);
+  if (!sportsVenue) {
+    throw new NotFoundException('Sports Venue with given ID not found');
   }
+
+  if (userType != 'owner') {
+    throw new ForbiddenException(
+      'Regular User is not allowed to delete this venue'
+    );
+  }
+
+  if (sportsVenue.ownerId.toString() !== ownerId.toString()) {
+    throw new UnauthorizedException(
+      'Sports Venue does not belong to the current user'
+    );
+  }
+
+  const deletedSportsVenue = await sportsVenueRepository.delete(id);
+  if (deletedSportsVenue == null) {
+    throw new NotFoundException('Sports Venue with given ID not found');
+  }
+
+  await publishSportsVenueDeletion({ sportsVenueId: id, ownerId });
+
+  return true;
 };
